+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - New #44: Add `pinChatMessage`, `unpinChatMessage` and `unpinAllChatMessages` methods.
 - New #46: Add `leaveChat` method.
 - New #47: Add `getChatAdministrators`, `getChatMemberCount` and `getChatMember` methods.
+- New #48: Add `InputSticker` type and `MaskPosition::toRequestArray()` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.